Clean up PR

+++ src/effects/effect_94.c
@@ -6,7 +6,7 @@ void func_E00BC1D0(void) {
 }
 
 f32 func_E00BC1D8(f32 arg0) {
-    return (f32) ((f64) (f32) shim_rand_int((s32) arg0) - ((f64) arg0 * 0.5));
+    return (f32)shim_rand_int((s32)arg0) - arg0 * 0.5;
 }
 
 INCLUDE_ASM(s32, "effects/effect_94", func_E00BC238);
